Implemented reverse delta unpack (#609)
* Change backup fetch order to reverse

* Change in empty directory check logic

- Move check for empty directory before deltaFetchRecursion call
- Remove check for empty directory from checkDbDirectoryForUnwrap

* Delete unwrapToEmptyDirectory and move checkDbDirectoryForUnwrap

unwrapToEmptyDirectory method is now obsolete because WAL-G checks for empty directory before deltaFetchRecursion. Also, the checkDbDirectoryForUnwrap function is now called on each unwrap call.

* Add new functions for increment and base files

-CreateFileFromIncrement creates empty local page file and fills it with the pages from the provided increment.
-RestoreMissingPages fills missing pages of local file with their base version.
-WritePagesFromIncrement writes pages from delta according to diffMap.

* Change in FileTarInterpreter's unwrap logic

This method now uses new functions for base and increment files manipulation and handles all other possible scenarios.

* Split the too long line

* Add comment to non-empty reader verification and move the verification to the separate function.

* Change the type of DatabasePageSize from uint16 to int64

* Fix the possible uint32 overflow of pageCount

* Split the logic into new and old for testing

* Fix pagefile_test

* Fix commands description & docs

* Add integration test for reverse unpack

* Fix possible false-positive error during unpack

+++ internal/configure.go
@@ -22,7 +22,7 @@ import (
 )
 
 const (
-	DefaultDataBurstRateLimit = 8 * int64(DatabasePageSize)
+	DefaultDataBurstRateLimit = 8 * DatabasePageSize
 	DefaultDataFolderPath     = "/tmp"
 	WaleFileHost              = "file://localhost"
 )
