+++ internal/incremental_page_reader.go
@@ -59,7 +59,7 @@ func (pageReader *IncrementalPageReader) AdvanceFileReader() error {
 	pageBytes := make([]byte, DatabasePageSize)
 	blockNo := pageReader.Blocks[0]
 	pageReader.Blocks = pageReader.Blocks[1:]
-	offset := int64(blockNo) * int64(DatabasePageSize)
+	offset := int64(blockNo) * DatabasePageSize
 	// TODO : possible race condition - page was deleted between blocks extraction and seek
 	_, err := pageReader.PagedFile.Seek(offset, io.SeekStart)
 	if err != nil {
@@ -84,7 +84,7 @@ func (pageReader *IncrementalPageReader) initialize(deltaBitmap *roaring.Bitmap)
 	headerBuffer.Write(IncrementFileHeader)
 	fileSize := pageReader.FileSize
 	headerBuffer.Write(utility.ToBytes(uint64(fileSize)))
-	pageReader.Blocks = make([]uint32, 0, fileSize/int64(DatabasePageSize))
+	pageReader.Blocks = make([]uint32, 0, fileSize/DatabasePageSize)
 
 	if deltaBitmap == nil {
 		err := pageReader.FullScanInitialize()
@@ -97,7 +97,7 @@ func (pageReader *IncrementalPageReader) initialize(deltaBitmap *roaring.Bitmap)
 
 	pageReader.WriteDiffMapToHeader(&headerBuffer)
 	pageReader.Next = headerBuffer.Bytes()
-	pageDataSize := int64(len(pageReader.Blocks)) * int64(DatabasePageSize)
+	pageDataSize := int64(len(pageReader.Blocks)) * DatabasePageSize
 	size = int64(headerBuffer.Len()) + pageDataSize
 	return
 }
@@ -106,7 +106,7 @@ func (pageReader *IncrementalPageReader) DeltaBitmapInitialize(deltaBitmap *roar
 	it := deltaBitmap.Iterator()
 	for it.HasNext() { // TODO : do something with file truncation during reading
 		blockNo := it.Next()
-		if pageReader.FileSize >= int64(blockNo+1)*int64(DatabasePageSize) { // whole block fits into file
+		if pageReader.FileSize >= int64(blockNo+1)*DatabasePageSize { // whole block fits into file
 			pageReader.Blocks = append(pageReader.Blocks, blockNo)
 		} else {
 			break
