+++ internal/pagefile.go
@@ -30,7 +30,7 @@ import (
 )
 
 const (
-	DatabasePageSize            = walparser.BlockSize
+	DatabasePageSize			= int64(walparser.BlockSize)
 	sizeofInt32                 = 4
 	sizeofInt16                 = 2
 	sizeofInt64                 = 8
